@@ -37,7 +37,7 @@ dependencies:
   go_router: ^15.1.1
   flutter_bloc: ^9.1.0
   flutter_dotenv: ^5.1.0
-  privy_flutter: ^0.0.8
+  privy_flutter: ^0.4.0
   convert: ^3.1.1
 
 dev_dependencies:
